+++ Sources/Servers/Project/ProjectMCPServer.swift
@@ -26,6 +26,7 @@ public enum ProjectToolName: String, CaseIterable, Sendable {
     case addTargetToTestPlan = "add_target_to_test_plan"
     case removeTargetFromTestPlan = "remove_target_from_test_plan"
     case setTestPlanTargetEnabled = "set_test_plan_target_enabled"
+    case setTestPlanSkippedTags = "set_test_plan_skipped_tags"
     case addTestPlanToScheme = "add_test_plan_to_scheme"
     case removeTestPlanFromScheme = "remove_test_plan_from_scheme"
     case listTestPlans = "list_test_plans"
@@ -139,6 +140,7 @@ public struct ProjectMCPServer: Sendable {
         let addTargetToTestPlanTool = AddTargetToTestPlanTool(pathUtility: pathUtility)
         let removeTargetFromTestPlanTool = RemoveTargetFromTestPlanTool(pathUtility: pathUtility)
         let setTestPlanTargetEnabledTool = SetTestPlanTargetEnabledTool(pathUtility: pathUtility)
+        let setTestPlanSkippedTagsTool = SetTestPlanSkippedTagsTool(pathUtility: pathUtility)
         let addTestPlanToSchemeTool = AddTestPlanToSchemeTool(pathUtility: pathUtility)
         let removeTestPlanFromSchemeTool = RemoveTestPlanFromSchemeTool(pathUtility: pathUtility)
         let listTestPlansTool = ListTestPlansTool(pathUtility: pathUtility)
@@ -209,6 +211,7 @@ public struct ProjectMCPServer: Sendable {
                 addTargetToTestPlanTool.tool(),
                 removeTargetFromTestPlanTool.tool(),
                 setTestPlanTargetEnabledTool.tool(),
+                setTestPlanSkippedTagsTool.tool(),
                 addTestPlanToSchemeTool.tool(),
                 removeTestPlanFromSchemeTool.tool(),
                 listTestPlansTool.tool(),
@@ -301,6 +304,8 @@ public struct ProjectMCPServer: Sendable {
                     return try removeTargetFromTestPlanTool.execute(arguments: arguments)
                 case .setTestPlanTargetEnabled:
                     return try setTestPlanTargetEnabledTool.execute(arguments: arguments)
+                case .setTestPlanSkippedTags:
+                    return try setTestPlanSkippedTagsTool.execute(arguments: arguments)
                 case .addTestPlanToScheme:
                     return try addTestPlanToSchemeTool.execute(arguments: arguments)
                 case .removeTestPlanFromScheme: